+++ NEWS.md
@@ -1,4 +1,4 @@
-# AzureVM 2.1.1.9000
+# AzureVM 2.2.0
 
 - Add ability to retrieve an SSH public key from an Azure resource object. See `?user_config` for more information.
 - New predefined configurations for VMs and VM scalesets: Ubuntu 20.04; Debian 10 (w/backports); Centos 8.1; RHEL 8.1 and 8.2; Windows Server 2019. All new configurations except Centos are available as both generation 1 and generation 2 VMs. See `?vm_config` and `?vmss_config` for more information.